+++ CHANGELOG.md
@@ -16,7 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - [JS] Initialize option `missingBehaviour` & `missingTranslationPrefix` with default values ([#398](https://github.com/fnando/i18n-js/pull/398))
 - [JS] Throw an error when `I18n.strftime()` takes an invalid date ([#383](https://github.com/fnando/i18n-js/pull/383))
-
+- [Ruby] Reset middleware cache on rails startup 
+([#402](https://github.com/fnando/i18n-js/pull/402))
 
 
 ## [3.0.0.rc12] - 2015-12-30